requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography>=3.4.7
requests==2.31.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
flask_sqlalchemy
